refactor: archiveView 에 CalendarView 관련 리팩토링
